backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
pillow==11.2.1
uvicorn==0.34.3
djoser==2.3.1
django-filter==25.1
psycopg2-binary==2.9.10
